changed the score in end game scene to double, and finished the func endGame with the transition and end scene

+++ DropGame/Scenes/EndGameScene.swift
@@ -10,7 +10,7 @@ import SpriteKit
 
 class EndGameScene : SKScene
 {
-    var score : Int = 0
+    var score : Double = 0.0
     
     //MARK: - Overriden Method
